Fix .data and .bss sections issue with unit tests
Migrate dummy Ram definitions to main_test.cpp in order to make sure
that it is defined for all builds.
