+++ library/L0_Platform/test/ram_test.cpp
@@ -4,49 +4,6 @@
 #include "L4_Testing/testing_frameworks.hpp"
 #include "L0_Platform/ram.hpp"
 
-namespace
-{
-struct DataSection_t
-{
-  int32_t a;
-  uint8_t b;
-  double d;
-  uint16_t s;
-};
-
-DataSection_t rom = {
-  .a = 15,
-  .b = 'C',
-  .d = 5.0,
-  .s = 12'346U,
-};
-
-DataSection_t ram;
-
-std::array<uint32_t, 128> bss_section;
-}  // namespace
-
-DataSectionTable_t data_section_table[] = {
-  DataSectionTable_t{
-      .rom_location = reinterpret_cast<uint32_t *>(&rom),
-      .ram_location = reinterpret_cast<uint32_t *>(&ram),
-      .length       = sizeof(rom),
-  },
-};
-
-// NOTE: Not used, but must be defined
-DataSectionTable_t data_section_table_end;
-
-BssSectionTable_t bss_section_table[] = {
-  BssSectionTable_t{
-      .ram_location = bss_section.data(),
-      .length       = bss_section.size(),
-  },
-};
-
-// NOTE: Not used, but must be defined
-BssSectionTable_t bss_section_table_end;
-
 namespace sjsu
 {
 TEST_CASE("Testing Ram Initialization")
@@ -54,33 +11,39 @@ TEST_CASE("Testing Ram Initialization")
   SECTION(".data")
   {
     // Setup
-    testing::ClearStructure(&ram);
+    memset(data_section_table[0].rom_location, 0, data_section_table[0].length);
 
     // Exercise
     sjsu::InitializeDataSection();
 
     // Verify
-    CHECK(memcmp(&rom, &ram, sizeof(rom)) == 0);
+    CHECK(memcmp(data_section_table[0].ram_location,
+                 data_section_table[0].rom_location,
+                 data_section_table[0].length) == 0);
   }
 
   SECTION(".bss")
   {
     // Setup
     // Setup: Fill the expected_blank_bss_section with all zeros.
-    std::array<uint32_t, bss_section.size()> expected_blank_bss_section;
+    std::array<uint32_t, 512> expected_blank_bss_section;
+
     memset(expected_blank_bss_section.data(), 0,
            expected_blank_bss_section.size());
-    // Setup: Fill the bss_section to be cleared with an arbitary value. This
+
+    // Setup: Fill the bss_section to be cleared with an arbitrary value. This
     //        value should be cleared to zero after calling
     //        `InitializeBssSection()`
-    memset(bss_section.data(), 0xA5A5A5A5, bss_section.size());
+    memset(bss_section_table[0].ram_location, 0xAA,
+           bss_section_table[0].length);
 
     // Exercise
     sjsu::InitializeBssSection();
 
     // Verify
-    CHECK(memcmp(expected_blank_bss_section.data(), bss_section.data(),
-                 bss_section.size()) == 0);
+    CHECK(memcmp(expected_blank_bss_section.data(),
+                 bss_section_table[0].ram_location,
+                 bss_section_table[0].length) == 0);
   }
 }
 }  // namespace sjsu
